added the option to check if the hermes2 already exist so it wouldn't fail

+++ MERCURY_Hamrony_License_Update.ps1
@@ -2,7 +2,7 @@
 $machineListFile = "C:\Users\iand\Documents\GitHub\MERCURY_Harmony_License_Server\machine_list.csv"
 
 # The file to look for on each machine
-$filePath = "C:\folder\file.dat"
+#$filePath = "C:\folder\file.dat"
 # The line to search for and the replacement text
 $searchText = "hermes"
 $replaceText = "hermes2"
@@ -24,12 +24,20 @@ foreach ($machine in $machines) {
             # Read the file content
             $fileContent = Get-Content -Path $remoteFilePath
 
-            # Replace the line of text
-            $fileContent = $fileContent -replace $searchText, $replaceText
+            # Converting fileContent to an array so we can use -contains to search for the word
+            $arrfileContent = -split $fileContent
 
-            # Write the modified content back to the file
-            Set-Content -Path $remoteFilePath -Value $fileContent
-            Write-Host "File modified on $machine."
+            # Find the text first and replce if it doesnt exists
+            if ($arrfileContent -contains $replaceText){
+                Write-Host "$replaceText already exist on the license file."
+            } else {
+                # Replace the line of text
+                $fileContent = $fileContent -replace $searchText, $replaceText
+
+                # Write the modified content back to the file
+                Set-Content -Path $remoteFilePath -Value $fileContent
+                Write-Host "File modified on $machine."
+            }
         } else {
             Write-Host "File not found on $machine."
         }
